Add: Weeks for task
Calculates how many weeks someone would need to accomplish a task over a given amount of days if they only had a certain amount of days per week. The program assumes one, two or three days per week are free for the task, and the percentage for the chance of one or three days is given as a CLI arg after -o/--one or -t/--three. The amount of days worth of work needed to complete the task is passed with -d/--days. If no arguments are passed the program assumes -o 0.25 -t 0.25 -d 20.